refactor(hvisor.ko): split scmi_server logic from hvisor.c
