+++ driver/Makefile
@@ -26,6 +26,10 @@ ifeq ($(ARCH), loongarch)
 	ccflags-y += -DLOONGARCH64
 endif
 
+hvisor-y := hvisor_main.o
+ifeq ($(VIRTIO_SCMI), y)
+	hvisor-y += scmi_server.o
+endif
 ivc-y := ivc_driver.o
 
 .SECONDARY: \
